move automapper config to new function, add try catch to StudentService and thread safe singleton
